Move internal hooks
