--- packages/react-client/src/hooks/internal.ts
@@ -1,1 +1,0 @@
-export { useFishjamClient_DO_NOT_USE } from "./useFishjamClient_DO_NOT_USE";
