+++ packages/react-client/src/hooks/internal/device/useDeviceApi.ts
@@ -1,8 +1,8 @@
 import { useMemo } from "react";
-import type { TrackManager } from "../../types/internal";
-import type { Device } from "../../types/public";
-import type { DeviceManager } from "../../DeviceManager";
-import { useDeviceManager } from "../deviceManagers/useDeviceManager";
+import type { TrackManager } from "../../../types/internal";
+import type { Device } from "../../../types/public";
+import type { DeviceManager } from "../../../DeviceManager";
+import { useDeviceManager } from "./useDeviceManager";
 
 type DeviceApiDependencies = {
   trackManager: TrackManager;
