+++ packages/react-client/src/hooks/internal/device/useDeviceManager.ts
@@ -1,6 +1,6 @@
 import { useCallback, useRef, useSyncExternalStore } from "react";
-import type { DeviceManagerState } from "../../types/internal";
-import type { DeviceManager, DeviceManagerEvents } from "../../DeviceManager";
+import type { DeviceManagerState } from "../../../types/internal";
+import type { DeviceManager, DeviceManagerEvents } from "../../../DeviceManager";
 
 const eventNames = [
   "deviceDisabled",
