+++ packages/react-client/src/hooks/internal/useFishjamClientState.ts
@@ -1,6 +1,6 @@
 import { useCallback, useMemo, useRef, useSyncExternalStore } from "react";
 import type { Component, MessageEvents, Peer, FishjamClient, GenericMetadata } from "@fishjam-cloud/ts-client";
-import type { PeerId } from "../types/internal";
+import type { PeerId } from "../../types/internal";
 
 const eventNames = [
   "socketClose",
