+++ packages/react-client/src/hooks/internal/useFishjamContext.ts
@@ -1,9 +1,9 @@
 import type { FishjamClient } from "@fishjam-cloud/ts-client";
 import { createContext, type MutableRefObject, useContext } from "react";
-import type { TrackManager } from "../types/internal";
-import type { DeviceManager } from "../DeviceManager";
+import type { TrackManager } from "../../types/internal";
+import type { DeviceManager } from "../../DeviceManager";
 import type { FishjamClientState } from "./useFishjamClientState";
-import type { BandwidthLimits, PeerStatus, ScreenshareApi } from "../types/public";
+import type { BandwidthLimits, PeerStatus, ScreenshareApi } from "../../types/public";
 
 export type FishjamContextType = {
   fishjamClientRef: MutableRefObject<FishjamClient>;
