+++ packages/react-client/src/hooks/useScreenShare.ts
@@ -3,7 +3,7 @@ import type { FishjamClient } from "@fishjam-cloud/ts-client";
 import { getRemoteOrLocalTrack } from "../utils/track";
 import type { TracksMiddleware, ScreenshareApi, PeerStatus } from "../types/public";
 import type { ScreenShareState } from "../types/internal";
-import { useFishjamContext } from "./useFishjamContext";
+import { useFishjamContext } from "./internal/useFishjamContext";
 
 interface ScreenShareManagerProps {
   fishjamClient: FishjamClient;
